Add extra tests
