+++ src/Objects/gameOptions.js
@@ -1,0 +1,15 @@
+const gameOptions = {
+  platformSpeedRange: [300, 300],
+  spawnRange: [80, 300],
+  platformSizeRange: [120, 300],
+  platformHeightRange: [-5, 5],
+  platformHeighScale: 20,
+  platformVerticalLimit: [0.4, 0.8],
+  playerGravity: 900,
+  jumpForce: 400,
+  playerStartPosition: 150,
+  jumps: 3,
+  coinPercent: 75,
+  spikePercent: 50,
+};
+export default gameOptions;
